requirements.txt
matplotlib~=3.7.1
tensorflow~=2.10.0
numpy~=1.23.5
keras~=2.10.0
pillow~=9.4.0